chore: publish koishijs/koishi-desktop@e7931cdea7aa8b08cb1c87f336ab5705a35fb2d8

+++ proc/koiproc.go
@@ -115,7 +115,7 @@ func (koiProc *KoiProc) Run() error {
 				if scnErr != nil {
 					l.Warn(fmt.Errorf("koiProc scanner err: %w", scnErr))
 				} else {
-					s := scn.Text() + "\n"
+					s := scn.Text()
 					out <- &s
 				}
 			}
@@ -138,6 +138,9 @@ func (koiProc *KoiProc) Run() error {
 //
 // This just sends the signal and do not wait for anything.
 func (koiProc *KoiProc) Stop() error {
+	if koiProc.cmd.Process == nil {
+		return nil
+	}
 	return koiProc.cmd.Process.Signal(syscall.SIGTERM)
 }
 
@@ -146,5 +149,8 @@ func (koiProc *KoiProc) Stop() error {
 // This just sends the signal and do not wait for anything.
 // If possible, use [KoiProc.Stop].
 func (koiProc *KoiProc) Kill() error {
+	if koiProc.cmd.Process == nil {
+		return nil
+	}
 	return koiProc.cmd.Process.Kill()
 }
